Complete Play handler, PlayRoulette. Add a global utility class for calculation and conversion

+++ app/src/main/java/com/robor/boule/MainActivity.java
@@ -20,7 +20,7 @@ public class MainActivity extends AppCompatActivity {
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
-        Game game = new Game(Game.GAMES.EURO_ROULETTE, new int[]{2,3,4}, new double[]{0,0});
+        Game game = new Game(Game.GAMES.EURO_ROULETTE, new int[]{2,3,4}, new int[]{0,0});
 
         FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
         fab.setOnClickListener(new View.OnClickListener() {
